blackbox: change units_to_bytes to size_representation
The units_to_bytes function is always failing due to the input
variable "size" always containing a period (e.g.: "10.00"); the
call fails with "ValueError: invalid literal for int() with base 10".

However, all of the callers of units_to_bytes() only use the return
value as a string in a dictionary, therefore returning the value
as a concatenation of size + units allows the tests to pass.

Since the units_to_bytes function now returns a string representation
of a size, rename the function to size_representation.

Signed-off-by: Bryan Gurney <bgurney@redhat.com>