Remove contradicting assertion in exec_command
Remove assertion that stderr is empty. This assertion is incompatible with
the rest of the method in two ways:
1. The method takes an expected exit code argument, which may not be 0.
If not 0, then there was an error and if stratis_cli writes to stderr when
there was an error, that is correct behavior. So, the assertion would cause
correct code to fail.
2. The method returns stdout and stderr, but asserts that stderr is always
the empty string and will fail if it is not. Since stderr is always the
empty string, if returned, there is no point in returning it. Removing
the assertion means that the returned stderr may be informative, since it is
not necessarily the empty string.

To see concretely how this might matter, examine PR
https://github.com/stratis-storage/stratis-cli/pull/332 and imagine the
behavior if the test test_no_args had not been removed when the bug
https://github.com/stratis-storage/stratis-cli/issues/248 was fixed.
If it had not been removed, then the right thing to do would have been to
remove the expectedFailure annotation (since the bug that it tested was
fixed). If the expectedFailure annotation had been removed and the test
had remained, it would have failed, since stratis-cli was fixed and does
print a usage member to stderr, as appropriate. This failure would have
been wrong, and we would have removed the assertion in that case, let's
remove it now.

Signed-off-by: mulhern <amulhern@redhat.com>