Double wait for D-Bus signals to 2 seconds
test_pool_add_different_data_after_cache is failing intermittently,
because for some reason the D-Bus signal caused by the addition of data
devices is not being emitted soon enough to catch.

Double the time that the stratisd_cert tests wait before sending the
kill signal, in hopes that this will resolve the problem.

Signed-off-by: mulhern <amulhern@redhat.com>