stratisd_cert.py: setUp -> tearDown
A notable effect of calling setUp instead of tearDown in the tearDown
method is that the dbus monitor script is cleaned up early.

That tearDown is not called is barely noticeable, since there is a
clean_up method that is invoked after the tests concludes, regardless of
whether it succeeded or failed.

Signed-off-by: mulhern <amulhern@redhat.com>